+++ src/modules/mandalart/components/mandalart-main-content.tsx
@@ -8,13 +8,9 @@ import { getCurrentUserName } from '@/shared/utils/get-current-user-name';
 import { useBatchUpdateTrigger } from '@/modules/mandalart/hooks/use-batch-update-trigger';
 import { useMandalartDataQuery } from '@/modules/mandalart/hooks/use-mandalart-data-query';
 import useFloatingSheetStore from '@/shared/hooks/use-floating-sheet-store';
-import { getBrowserClient } from '@/shared/utils/supabase/browser-client';
 import { useQueryClient } from '@tanstack/react-query';
 import { useEffect } from 'react';
 import { createTodoListkey } from '@/modules/mandalart/services/create-todo-list-key';
-import { QUERY_KEY } from '@/shared/constants/query-key';
-import { TodoPayloadType } from '@/modules/mandalart/types/realtime-type';
-import { useBroadcastStore } from '@/modules/mandalart/hooks/use-broadcast-store';
 import Spacer from '@/components/commons/spacer';
 import Title from '@/components/commons/title';
 import Text from '@/components/commons/text';
@@ -27,6 +23,7 @@ import { AvatarStack } from './mandalart-avatar-stack';
 import { useUsersStore } from '../hooks/use-users-store';
 import { getCurrentUserId } from '@/shared/utils/get-current-user-id';
 import Button from '@/components/commons/button';
+import { useRealtimeBroadCastRoom } from '../hooks/use-realtime-broadcast-room';
 
 /**
  * Memo: useCurrentUserName 훅으로 닉네임을 가져와서
@@ -42,82 +39,28 @@ const MandalartMainContent = ({
   user,
   mandalartId,
 }: MandalartMainContentProps) => {
-  const supabase = getBrowserClient();
   const queryClient = useQueryClient();
-
-  useBatchUpdateTrigger();
-  useRealtimePresenceRoom('avatar-room', user);
+  const { data, isPending, isError } = useMandalartDataQuery(mandalartId);
 
   const isVisible = useFloatingSheetStore((state) => state.isVisible);
   const currentUsers = useUsersStore((state) => state.currentUsers);
 
-  const { data, isPending, isError } = useMandalartDataQuery(mandalartId);
+  useRealtimePresenceRoom(`avatar-room ${mandalartId}`, user);
+  const broadcastChannel = useRealtimeBroadCastRoom(
+    `broadcastChannel ${mandalartId}`
+  );
+  useBatchUpdateTrigger();
 
   useEffect(() => {
     if (isPending) return;
     createTodoListkey(queryClient, data);
   }, [isPending, data, queryClient]);
-  const addBroadcastStore = useBroadcastStore(
-    (state) => state.addBroadcastStore
-  );
+
+  if (isPending || !broadcastChannel) return <div>Loading...</div>;
+  if (isError) return <div>error</div>;
 
   const username = getCurrentUserName(user);
   const userId = getCurrentUserId(user);
-
-  //TODO 룸네임받는 훅으로 수정
-  const broadcastChannel = supabase.channel('broadcastChannel');
-  broadcastChannel
-    .on('broadcast', { event: 'shout' }, (payload) => {
-      addBroadcastStore(payload.payload);
-      //투두일때
-      if ('action' in payload.payload) {
-        if (payload.payload.action === 'UPDATE') {
-          queryClient.setQueryData(
-            QUERY_KEY.todolist(payload.payload.cell_id),
-            (todoList: TodoPayloadType[]) => {
-              return todoList.map((item) =>
-                item.id === payload.payload.id ? payload.payload : item
-              );
-            }
-          );
-
-          queryClient.setQueryData(
-            QUERY_KEY.todo(payload.payload.id),
-            payload.payload
-          );
-          return;
-        }
-
-        if (payload.payload.action === 'CREATE') {
-          queryClient.setQueryData(
-            QUERY_KEY.todolist(payload.payload.cell_id),
-            (todoList: TodoPayloadType[]) => {
-              return [...todoList, payload.payload];
-            }
-          );
-          return;
-        }
-
-        if (payload.payload.action === 'DELETE') {
-          queryClient.setQueryData(
-            QUERY_KEY.todolist(payload.payload.cell_id),
-            (todoList: TodoPayloadType[]) => {
-              return todoList.filter((item) => item.id !== payload.payload.id);
-            }
-          );
-          return;
-        }
-      }
-
-      queryClient.setQueryData(
-        [payload.payload.category, payload.payload.id],
-        payload.payload.value
-      );
-    })
-    .subscribe();
-
-  if (isPending) return <div>Loading...</div>;
-  if (isError) return <div>error</div>;
 
   return (
     <div className='flex flex-col items-center'>
